requirements.txt
python-dotenv~=1.0.0
aiogram~=3.3.0
aiosqlite~=0.19.0
validators~=0.22.0
python-okx~=0.4.1
requests~=2.32.5